- Finished Basic Rendering -- Added Camera Component -- Added Shaders & ShaderPrograms - Added Kyeboard & Mouse Input - Switched to using JOML

+++ src/main/java/org/daedalus/graphics/Resolution.java
@@ -12,7 +12,7 @@ public class Resolution{
     }
     
     public float getAspectRatio(){
-        return width / height;
+        return (float)width / (float)height;
     }
     
 }
